git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@14674 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickWand/drawing-wand.c
@@ -5542,7 +5542,7 @@ WandExport MagickBooleanType DrawSetStrokeDashArray(DrawingWand *wand,
                 wand->name);
               return(MagickFalse);
             }
-          for (i=0; i < n_new; i++)
+          for (i=0; i < (ssize_t) n_new; i++)
           {
             CurrentContext->dash_pattern[i]=0.0;
             if (dasharray != (double *) NULL)
@@ -5556,7 +5556,7 @@ WandExport MagickBooleanType DrawSetStrokeDashArray(DrawingWand *wand,
       else
         if (dasharray != (double *) NULL)
           {
-            for (i=0; i < n_new; i++)
+            for (i=0; i < (ssize_t) n_new; i++)
             {
               if (i != 0)
                 (void) MvgPrintf(wand,",");
